Stroke -> SatinStitch + RunningStitch, add ManualStitch
